update announcement on top of mobile version
